Started getting socket comms up and running
* Added a Process class that now takes care of socket comms.
* Service class now extends Process class.
* Created a Container class for short lived functions created by runDelay and such.

NOTE: There is a windows hack to allow socket_create() to work in the child program.  We needed to copy the $_SERVER array into proc_open() as the environment for this to all work.  Works fine though and we should now be able to communicate.

Just need to implement websockets in Process so that it can talk to the server.