Event trigger default behaviour changed
Previously, the triggering client would receive the event. ie: the event is echo'd back to them.  This is now disabled by default and the client will not receive the event if they are subscribed to it.  If they are not subscribed, nothing will change.

The trigger packet can now contain "echo=true" which will cause the event to be echo'd ONLY IF the client has subscribed to this event.

MVC Client has been updated to send this parameter.