Improved error and exception handling
The server already had decent error and exception handling under certain circumstances.  However, this was being interferred with by the inclusion of the Hazaar\Application source file which sets up application error/exception handling. 

As it turns out, the server did not really require the application class to be loaded and it only used it to determine the runtime path.  Because of this I have created a Server::runtimePath() method on the Warlock server class itself so that the Hazaar\Application class is not loaded at all.

This caused a problem with helper functions.  So now I have also made it load the Hazaar\Loader class directly.  I then use reflection to get the file it is declared in and add that as an include path.  Then I'm free to load the helper file directly.  This turns out to use a little bit less memory, but may increase startup time (big deal).