Ditched the wait OK response sending a TRIGGER
This is pretty much the root cause of issues outlined in #18.  The weirdness came from there being an EVENT packet in the receive buffer before sending a TRIGGER.  So waiting for an OK at that point would return the EVENT which was then interpreted as a "not OK" result.

fixes #18