Fixed up the schedule action processor
There was a bit of an issue which caused some sceduled actions to never actually execute.  Basically, if the order in which the actions was defined was had an action that would execute before the action defined before it, it would never execute!  This was simply due to an elseif being used instead of an if.