Dynamic services are up and running!
This is the first iteration of this feature.  I have services starting up and shutting own either manually (with a kill message) or automatically when the parent client disconnects.

I feel it's a bit messy though, so I'm going to tweak a few things now that I have it running with a nice test app.

refs #21