Fixed dynamic service cancelling when spawned from backend
When spawning from the backend we use a Hazaar\Warlock\Control class which connects to the server, sends a message and then disconnects.  The problem is that dynamic services are tied to the client.  This is not required for backend spawns and so we now "detach" the service so it will continue to run when the backend disconnects.