Remove "dynamic" service config option
Now that we require a dynamic service to have a runOnce() method to operate, there is no point having a dynamic config option.  This was supposed to allow us to enable the service, but not actually run it unless called.  Changes to spawn have now made this option redundant and a service can be spawned dynamically if it has a runOnce() method.