Added config load check
If the config file is not loaded we now consider Warlock to be disabled and throw an exception and exit.  This is tested both in Control and Server so that unles warlock is configured, nothing will work because without a config file, nothing will work anyway but it's harder to know what's actually gone wrong.