Major rework of the master server
This is a major update to the Warlock server.  All client specific code has been moved into the client class so that client objects will handle all the client communications.  If a client wants to do something "global" like trigger an event then it executes this on the server object itself.  A client can register for admin access to execute commands directly on the server.  This adds some security to the admin stuff.  All WebSocket comms works.  Service enable/disable, spawn/signal/kill and all that, have been tested and operate as before.  Things should just be a bit more stable.