Stage one completed.  We now have KV store!
Pretty easy to implement actually.  The protocol and comms systems already set up for Warlock comms made this very easy.  

At the moment we just have the basics done.  So GET, SET, HAS, DEL, LIST and CLEAR.  There is no support for namespaces or timeouts as yet.  That is coming next.