Added instance tracking to Warlock control
We now track Control instances based on server/port.  Using Control::getInstance() takes care of all of this transparently as it did before.  However it allows multiple instances to be used that connect to different servers and/or ports.  This was required because we now use Control in the warlock cache backend which can connect to anything.