Added option to start control without connection
Everywhere else has a socket connection check, so it's pretty much OK to carry on without a connection.  The issue is the connection timeout, so if a connection is not made we would throw an exception.  Mainly because there is no point having a control object if there is no connection.  however I have come across a situation where I just want to carry on if there is no connection to Warlock, such as during testing or when we don't want a Warlock failure to hold up other things.  Because of this it is now possible to indicate when instantiating the control object whether a connection is required.  The default is TRUE for backwards compatibility.  Setting this new argument to false will still attempt to connect but if that fails it will carry on as normal.