Background exec support for static methods
The first part of being able to have serviceless code execution and executing scheduled code in it's own process, configured in the main config and without a service class, is to be able to schedule execution that references static class methods.  This commit allows this to work from the standard `Warlock\Control::schedule()` and `Warlock\Control::runDelay()` methods by instead of supplying a closure, and using any 'callable' object, such as an array.  see `is_callable()`.