Added global event queue
It's now possible to subscribe to global events.  These are events that have no service and are executed with ad-hoc jobs.  This is almost exactly the same as using runDelay or Schedule to execute a background job except the job starts when a trigger is called and the arguments are the same as a normal event handler where the first argument is the trigger data and the second is the whole payload.