Services now using pipes for bi-directional comms
Finally got this done.  The pipe receiver is outside the main select loop so there can be some lag in receiving messages from services, but I will fix this later.  Right now, the receiver is called and allows a process to handle incoming data from the pipe instead of a socket.  This means that there is now bi-directional communication between services and Warlock master.

Along with this, I have created the concept of "connections" and there is currently two types of connections.  A socket and a pipe.  Services use pipes for comms at the moment, and Control uses a socket.  However to solve #49 it should be possible to get a service to choose what it wants to use, based on config settings.  This means we can run a service and connect via a socket.