Unified select for all streams (pipes and sockets)
So everything was working, but there was possible lag on receiving data from a service because the processing of received data happened outside the main select loop.  I have switched over to using socket_streams so that I can now use stream_select on the main select loop.  This then allowed me to add the pipe streams to the sockets array and track them the same way.  All data communications, be it socket or stream, is now instantaneous.  Yay!