Moved dependencies into Warlock library
There was really only tw things that needed to happen here.

* Hazaar\Application::runStdin() was moved into the Warlock library.  It was added as Hazaar\Warlock\Process::runner() and was essentially unchanged.  Only changes were that '$this', which was the Application, needed to be provided as an argument instead, which was done so from Runner.php.
* Hazaar\Application\Protocol was also moved into Warlock as Hazaar\Warlock\Protocol and all references updated.

Because these two dependencies are now in Warlock, their existence in Core is neither required nor a concern.  If they are still there, they are simply not used.

This means that going forward, Warlock 3 will be backwards compatible with pretty much any version of Hazaar MVC core, however I am limiting this to Core 2.1 which is pretty old but still the earliest known release on packagist.