Initial working of remote services
This has a remote service working.  It can be started manually and then connects back to Warlock using sockets instead of pipes (the old way).  This was quite simple due to the work I did previously in creating connection classes for processes, which meant that I can now check if the service is running in remote mode and use a socket instead of a pipe.

While the service works, it isn't registering as a service and only has typical user rights at the moment.