Peer connections are working
First stage of peer connections is working.  I have two instances of Warlock connecting to each other and communicating in a limited fashion.  The full websocket protocol is not yet available.  I think that I have duplicated some work so I'm committing now so that I can look into extending the Server\Peer class from the Server\Client class so that websockets protocol is baked in.