Converted Warlock\Peer to extend Warlock\Client
While I was writing the Peer class, I noticed that I was about to start duplicating a lot of code that is already proven reliable in the Client class.  So I have extended peer of the back of client, added the outbound connection bits and then just overridden the processCommand method to handle commands in it's own way.  Seems to be working and will use the same frame/packet buffering that a client has, making it just as reliable.

As I write this, I have two instances connected, online and sending WebSocket pings back and forth.  :D