Switched to ASYNC peer connections
This allows us to maintain responsiveness while trying to bring the connection back up.  To say, getting this to work was a pain in the ass, is an understatement.  There is no "nice" way to be notified or detect that the socket is connected, so we just give it one iteration to connection  (should be 1 second) and if it's not up, we tear it down and try to connect again.