Major updates and code shuffle
Work done so far:
* Moved signal handling into it's own class
* Cluster handles the clients/peers nodes.
* Master is now only responsible for connections, main loop and jobs, but will be losing jobs later because I will need to move that into the cluster.  I will create it's own class just as I did for signals.

Current status is I have client connections and signals working again.  Jobs/services/KV/etc are all currently offline while I re-write most things.  I have also just gotten peer connections working, however the receiver peer things a client is connecting so sends the init frame, which causes an error on the sender peer, so it sends back an error frame, which then trips an error on the receiver so it sends an error frame, and the endless loop begins.  ;)

I will make it handle this better just in case, but this will not happen under good circumstances as peers do not need the init frame.