Improved connection handling
This update has better connection handling.  Now if peer connects are lost we will try and reconnect.  We also monitor the connection timeout ourselves.  This can be set in the config.  I have also implemented a better handshake process on both connections and peers, as we now use a status indicator.