Updated the cluster peer instantiation
Peers were created on construction of the cluster object.  Now we create them when the cluster is started.  This makes for some nicer logging, but allows us to handle duplicate connections a bit more nicely.  

So yeah, we also filter out duplicate active connections.  This is based on host:port only but will allow connections with the same host:port to be enabled/disabled with different settings to allow for development/experimentation.