Added extended global events
If a global event method is stored in a Service, it now no longer has to be a static method.  Instead, the Service will be instantiated directly.  To solve the problem of when init() should be called, I have added a new 'construct' function that can optionally be added.  This can be used to initialise things that don't require a running process.