Fixed buffer + performance enhancements
The split frame buffer was broken.  This meant that if a frame was actually split over multiple buffer reads (which seems to be at 8192 bytes) then unprocessed buffer data was discarded.  We now hang on to it in a more efficient manner.  Testing with large packets of 50MB has resulted in success. 

NOTE:  I tried a 100MB packet and it worked once, but mostly failed to send due to the fwrite() buffer getting full.  I could setup an output/write buffer internally as well, but right now there's not really any reason to send anything so large.  Maybe fix that one later.